Client : Refactor SettingsNavigation.vue. Repositioned
