+++ client/src/components/views/seb-server/settings-navigation/widgets/SettingsNavigation.vue
